pyproject.toml
[project]
name = "deepsearch-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=0.9.9",
    "openai==1.98.0",
    "openai-agents>=0.2.5",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
]
